+++ src/db/db.json
@@ -328,21 +328,21 @@
         "cost": 10
     },       
     {
-        "id": 2,
+        "id": 3,
         "name": "chocolate",
         "description": "Grocery items",
         "cost": 5
        
     },
     {
-        "id": 2,
+        "id": 4,
         "name": "T-shirt",
         "description": "wear",
         "cost": 25
         
     },
     {
-        "id": 2,
+        "id": 5,
         "name": "sneaker",
         "description": "foot wear",
         "cost": 35
